clobot-adapter-main/requirements.txt
annotated-types==0.5.0
asyncqt==0.8.0
paho-mqtt==1.6.1
pydantic==2.0.2
pydantic_core==2.1.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dotenv==1.0.0
Quamash==0.6.1
typing_extensions==4.7.1
